Adjust sourced items on actors to use remapped UUIDs
Moves the `parseUuid` changes from `JournalSpellListPageSheet` into
a shared implementation and applied it to `SourcedItemsMap` during
get and set operations so it always handles remapped UUIDs. Because
the initial setup of sourced items occurs before the UUID redirects
are loaded, the system will re-adjust the map during the `ready`
hook.
